+++ crates/draco-core/src/mesh_edgebreaker_traversal_valence_decoder.rs
@@ -204,7 +204,7 @@ impl<'a> MeshEdgebreakerTraversalValenceDecoder<'a> {
                     in_buffer,
                     &mut self.context_symbols[i],
                 )
-                .map_err(|err| DracoError::general(format!("Valence context {i} of 6: {err}")))?;
+                .map_err(|err| err.context(format!("Valence context {i} of 6")))?;
                 // Set counter to read from back
                 self.context_counters[i] = num_symbols as i32;
             }